[KM-323] 수정:인증이 필요한 API들에 대해 컨트롤러 분리
